Add ExecWithLua to LuaMessage and move to lua/types.rs
